requirements.txt
# Core LangChain / LangGraph
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langgraph>=0.1.0

# HTTP client
httpx>=0.27.0

# gRPC
grpcio>=1.63.0
grpcio-tools>=1.63.0
protobuf>=5.27.0

# OpenAPI / YAML parsing
pyyaml>=6.0.1
jsonschema>=4.22.0
openapi-schema-validator>=0.6.0

# Report generation
jinja2>=3.1.4
markdown>=3.6

# Utilities
python-dotenv>=1.0.1
rich>=13.7.1
anyio>=4.4.0
